+++ crates/oxc_transformer/src/es2022/class_properties/private_field.rs
